Create MaxValidPointsInACircle
Challenge:  You are given a string of "tags", S, associated with two integer arrays corresponding to x and y coordinates, X and Y respectively.  Suppose a circle is drawn around the origin in such a way that the circle contains the maximum number of tagged points with NO repeating tags inside of the circle (i.e. the circle must contain the maximum number of unique tags so that no two of the same tags lie inside or on the edge of the circle).  Create a program that will give the correct number of valid points for any size array (Note: len(X) = len(Y) = len(S)).