Create BankCypherID
Challenge: A bank wants to cipher their customers' ID in such a way that the new ID is the first element in the old ID, then the last element in the old, then the second element, then the second to last, , then the third, and so on (i.e. "123456" becomes "162534" and "abcde" becomes "aebdc").  Create a program that when given a string, A, of any size will output the correct cipher.