Create RemoveIslandsIn2DMatrix
You are given a 2-dimensional array of 0's and 1's.  The task is to only have 1's that are either on the border or connected to a series of 1's that leads back to the border (think of the matrix as a graph).  Any 1's that are not on the border or connected to a series of 1's that is on the border must be removed and replaced with a 0 (all of the "islands" must be removed and changed to 0).

Example input:
[[1, 0, 0, 1, 0]
[0, 1, 0, 1, 1]
[0, 1, 0, 1, 0]
[1, 0, 1, 0, 1]
[0, 1, 0, 0, 0]]

Expected output:
[[1, 0, 0, 1, 0]
[0, 0, 0, 1, 1]
[0, 0, 0, 1, 0]
[1, 0, 0, 0, 1]
[0, 1, 0, 0, 0]]